Move periodic screen scanning to Pokefly service
One we switch out of the main app screen (MainActivity), the activity
shouldn't really be doing any background work. Services are meant for
doing background work. So, move the periodic screen scan to the Pokefly
service.
